backend/requirements.txt
chromadb==1.0.20
python-dotenv==1.1.1
fastapi[standard]==0.116.1
langchain
langchain-core
langchain-chroma
langchain-cohere
langchain_google_genai
typing
pydantic
cohere